requirements.txt
asgiref==3.6.0
Django==4.2
django-cleanup==7.0.0
django-cors-headers==3.14.0
django-environ==0.10.0
djangorestframework==3.14.0
Pillow==9.5.0
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.4
ua-parser==0.16.1
user-agents==2.2.0
